+++ src/components/Forms/FormRegister.scss
@@ -58,3 +58,4 @@
 	gap: 4px;
 	align-self: center;
 }
+
