+++ src/components/Popup/SuccessEmailed.scss
@@ -1,0 +1,31 @@
+.popup__text {
+	margin: 0;
+	padding-top: 12px;
+	padding-bottom: 28px;
+	color: var(--Primary-Black-light, #515151);
+	font-family: Roboto;
+	font-size: 16px;
+	font-style: normal;
+	font-weight: 400;
+	line-height: 20px;
+	display: flex;
+	align-items: start;
+	text-align: justify;
+}
+
+.register-form__button-register {
+	margin: 0;
+	border-radius: 8px;
+	background: var(--Accent-Blue, #1478ff);
+	padding: 0;
+	width: 326px;
+	height: 44px;
+	color: var(--Primary-White, #fcfcfc);
+	font-family: Roboto;
+	font-size: 14px;
+	font-style: normal;
+	font-weight: 700;
+	line-height: 20px;
+	text-align: center;
+	align-self: center;
+}
