+++ src/utils/api.js
@@ -8,7 +8,9 @@ function checkResponse(res) {
 			if (res.ok) {
 				return data;
 			}
-			return Promise.reject(new Error(data.detail || `Ошибка: ${res.status}`));
+			return Promise.reject(
+				new Error(Object.values(data) || `Ошибка: ${res.status}`)
+			);
 		});
 	}
 	return Promise.reject(new Error(`Ошибка: ${res.status}`));
